+++ src/main/java/com/cognizant/agriserve/service/AdvisoryContentService.java
@@ -6,5 +6,5 @@ import java.util.List;
 public interface AdvisoryContentService {
     AdvisoryContent saveContent(AdvisoryContent content);
     List<AdvisoryContent> getAllActiveContent();
-    void softDeleteContent(Integer id);
+    void softDeleteContent(Long id);
 }
